requirements.txt
python-dotenv
fastapi
uvicorn
langchain
langchain-community
langchain-openai
huggingface_hub